Library 2023/DocumentsAPI/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
ecdsa==0.17.0
fastapi==0.73.0
h11==0.13.0
idna==3.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.5
